feat: 상품 상세 조회 시 랭크를 포함

+++ apps/commerce-api/src/main/java/com/loopers/domain/ranking/RankingQueryResult.java
@@ -4,7 +4,7 @@ public record RankingQueryResult() {
 
     public record SearchRanks(
             Long productId,
-            Integer rank
+            Long rank
     ) {
     }
